+++ src/Defaults/IceHawkDelegate.php
@@ -1,0 +1,27 @@
+<?php
+/**
+ * @author hollodotme
+ */
+
+namespace Fortuneglobe\IceHawk\Defaults;
+
+use Fortuneglobe\IceHawk\Interfaces\SetsUpEnvironment;
+
+/**
+ * Class IceHawkDelegate
+ * @package Fortuneglobe\IceHawk\Defaults
+ */
+class IceHawkDelegate implements SetsUpEnvironment
+{
+	public function setUpErrorHandling()
+	{
+	}
+
+	public function setUpSessionHandling()
+	{
+	}
+
+	public function setUpGlobalVars()
+	{
+	}
+}