+++ tests/Unit/Events/IceHawkEventListenerTest.php
@@ -5,9 +5,9 @@
 
 namespace Fortuneglobe\IceHawk\Tests\Unit\Events;
 
-use Fortuneglobe\IceHawk\AbstractEventListener;
 use Fortuneglobe\IceHawk\Events\HandlingRequestEvent;
 use Fortuneglobe\IceHawk\Events\IceHawkWasInitializedEvent;
+use Fortuneglobe\IceHawk\PubSub\AbstractEventSubscriber;
 use Fortuneglobe\IceHawk\RequestInfo;
 use Fortuneglobe\IceHawk\Requests\GetRequest;
 
@@ -20,7 +20,7 @@ class IceHawkEventListenerTest extends \PHPUnit_Framework_TestCase
 		$getRequest    = new GetRequest( $requestInfo, [ ] );
 		$handlingEvent = new HandlingRequestEvent( $getRequest );
 
-		$mock = $this->getMockBuilder( AbstractEventListener::class )
+		$mock = $this->getMockBuilder( AbstractEventSubscriber::class )
 		             ->setMethods( [ 'getAcceptedEvents' ] )
 		             ->getMockForAbstractClass();
 
@@ -28,24 +28,24 @@ class IceHawkEventListenerTest extends \PHPUnit_Framework_TestCase
 		     ->method( 'getAcceptedEvents' )
 		     ->willReturn( [ IceHawkWasInitializedEvent::class ] );
 
-		/** @var AbstractEventListener $mock */
+		/** @var AbstractEventSubscriber $mock */
 		$this->assertTrue( $mock->acceptsEvent( $initEvent ) );
 		$this->assertFalse( $mock->acceptsEvent( $handlingEvent ) );
 	}
 
 	/**
-	 * @expectedException \Fortuneglobe\IceHawk\Exceptions\EventListenerMethodNotCallable
+	 * @expectedException \Fortuneglobe\IceHawk\PubSub\Exceptions\EventSubscriberMethodNotCallable
 	 */
 	public function testNotImplementedListenerMethodThrowsException()
 	{
 		$requestInfo = RequestInfo::fromEnv();
 		$initEvent   = new IceHawkWasInitializedEvent( $requestInfo );
 
-		$mock = $this->getMockBuilder( AbstractEventListener::class )
+		$mock = $this->getMockBuilder( AbstractEventSubscriber::class )
 		             ->setMethods( [ 'getAcceptedEvents' ] )
 		             ->getMockForAbstractClass();
 
-		/** @var AbstractEventListener $mock */
+		/** @var AbstractEventSubscriber $mock */
 		$mock->notify( $initEvent );
 	}
 
@@ -54,13 +54,13 @@ class IceHawkEventListenerTest extends \PHPUnit_Framework_TestCase
 		$requestInfo = RequestInfo::fromEnv();
 		$initEvent   = new IceHawkWasInitializedEvent( $requestInfo );
 
-		$mock = $this->getMockBuilder( AbstractEventListener::class )
+		$mock = $this->getMockBuilder( AbstractEventSubscriber::class )
 		             ->setMethods( [ 'getAcceptedEvents', 'whenIceHawkWasInitialized' ] )
 		             ->getMockForAbstractClass();
 
 		$mock->expects( $this->once() )->method( 'whenIceHawkWasInitialized' )->with( $initEvent );
 
-		/** @var AbstractEventListener $mock */
+		/** @var AbstractEventSubscriber $mock */
 		$mock->notify( $initEvent );
 	}
 }